refactor(orchestrator): split worker-outcome.ts into focused modules
Decompose the 434-line monolithic worker-outcome.ts into 7 focused
modules under src/orchestrator/worker-outcome/:

- types.ts: shared types (WorkerOutcomeInput, PreparedWorkerOutcome, issueRef, outcomeToStatus)
- prepare.ts: flush persistence, remove running entry, fetch latest issue, persist attempt snapshot
- completion-writeback.ts: tracker state transition + completion comment (non-blocking)
- terminal-paths.ts: 5 exit-path handlers (service-stopped, terminal cleanup, inactive, operator abort, cancelled/hard failure)
- retry-paths.ts: 4 retry handlers (continuation, exhausted, error backoff, model override)
- stop-signal.ts: DONE/BLOCKED flow (git post-run, attempt patch, claim semantics)
- index.ts: thin facade preserving the public handleWorkerOutcome() signature

Zero behavior change — all 1554 tests pass. Branch invariants preserved:
operator_abort does NOT release claim, DONE claim is sticky, BLOCKED releases.
